Changes to platform/security code

+++ drivers/platform/msm/pft.c
@@ -63,6 +63,7 @@
 #include <linux/fdtable.h>
 #include <linux/selinux.h>
 #include <linux/security.h>
+#include <linux/lsm_hooks.h>
 
 #include <linux/pft.h>
 #include <uapi/linux/msm_pft.h>
@@ -914,7 +915,7 @@ int pft_inode_post_create(struct inode *dir, struct dentry *dentry,
 	case PFT_STATE_KEY_LOADED:
 		/* Check whether the new file should be encrypted */
 		if (pft_is_current_process_registered()) {
-			u32 key_index = pft_get_app_key_index(current_uid());
+			u32 key_index = pft_get_app_key_index((u32)current_uid());
 			ret = pft_tag_file(dentry, key_index);
 			if (ret == 0)
 				pr_debug("key loaded, pid [%u] uid [%d] is creating file %s\n",
@@ -1497,7 +1498,7 @@ static int pft_set_inplace_file(struct pft_command *command, int size)
 	pft_sync_file(filp);
 
 	rc = pft_tag_file(pft_dev->inplace_file->f_dentry,
-			  pft_get_app_key_index(current_uid()));
+			  pft_get_app_key_index((u32)current_uid()));
 
 	if (!rc) {
 		pr_debug("tagged file %s to be encrypted.\n",
